sqlproxyccl: return an error if resumeForwarding fails
If resumeForwarding() returned an error, the error was logged, but
TransferConnection returned nil. I ran into this while trying to figure
out while a TransferConnection test was failing. The fact the error is
hidden in a defer was extra difficult to debug, because a debugger steps
over defer functions unless a break point is placed within the defer.

Release note: None