instancestorage: remove commit deadline
Previously, instancestorage used the session expiration as a commit
deadline when claiming an instance ID. If the passed in expiration was
reached before an instance ID was claimed, CreateInstance would get
stuck in an infinite loop.

The expiration argument is unecessary. Users of instance ID must ensure
the session t is valid, but it is safe to claim an instance ID with an
expired session.

Release note: None