kvcoord: route to any replica if no leaseholder info is known
Today if we do not know about the ClosedTimestampPolicy and we do not know
about the current leaseholder, we'll route the request to the nearest *voting*
replica. This replica may be in a remote region. Not knowing these things is
the common case when starting up a server. Non-voting replicas generally know
just as well as voting replicas where a leaseholder might be. As importantly,
as we transition system tables to being global, this is our opportunity to
have the request served by a follower.

Release note: None