tenantcostmodel: charge for cross region network traffic
Previously the 'tenant_cost_model.region_multiplier_table' allowed
configuring the cost of multi-region operations by applying a multiplier
to the entire operation cost. There were three problems with this approach
that prevented deploying it to production:

1. There was a flat multiplier for reads and writes. Because the fixed
	 overhead of a read is much lower, the cost impact of cross region
	 network traffic is much larger when compared to writes. The break
	 even read multiplier was an order of magnitude higher than the break
	 even write multiplier.
2. The correct multiplier depends heavily on how many bytes are in an
	 operation. Large operations require a larger multiplier to break even
	 than smaller operations. Again the break even multiplier for large
	 operations needed to be an order of magnitude larger than the break
	 even for small operations.
3. The original setting assumed the cost of network traffic was
	 symmetric between two different regions. That assumption is true for
	 some cloud providers, but is not true for all cloud providers.

Now, the 'tenant_cost_model.network_cost' setting allows configuring the
RU/byte cost of network traffic between region pairs. The cost is
computed by the dist sender because computing the network cost requires
locality information for the local node, the target replica, and all
replica's in the range.

The network cost only applies to logical byte traffic, it does not apply
to RPC overhead. This is believed to be okay, because the fixed RU cost
of each request and batch is high enough to cover the network cost of
the operation overhead.

This only accounts for network traffic incurred by reads or writes
between the sql server or the kv server. It does not account for network
traffic generated by zone configuration changes, replica rebalancing, or
distsql.
- Currently distsql is disabled in production Serverless and will only
	be reenabled once distsql properly accounts for network traffic.
	Distsql should use the same byte per RU cost table as sql server <->
	kv traffic.
- Zone configuration is expected to generate a small enough amount of
	the overall network traffic that we do not need to charge for it. We
	can revise this decision in the future if production data shows that
	is not true.
- Replica rebalancing is something controlled by CRDB. If rebalancing
	cost ends up being a problem, we could look into reducing the amount
	of cross-region rebalancing traffic or we could start charging for
	rebalancing traffic.

This change is backwards incompatible and will be back ported to 23.1.
This is safe because Serverless is the only user of the setting.

Release Note: none

Part of: https://cockroachlabs.atlassian.net/browse/CC-7140