cloud: randomize name of buckets that do not exist
Previously, we used a hard coded bucket name when tested the GCS and S3
behavior of buckets that do not exist. This made it possible for people
to create that bucket and break the test. Now, the test uses a random
bucket name.

Release Justification: Test Only Change
Release Note: None
Fixes: #135307
Fixes: #135348
Fixes: #135349
Fixes: #135350
Fixes: #135351
Fixes: #135352