logical: use destination descriptor for decoding refresh
Previously, when a cput failed, the source descriptor was used to decode
the returned KV. That only works if the source descriptor has the same
physical encoding. This is usually true for LDR, but the encodings will
differ if the tables have different column ids.

Now, a descriptor from the destination cluster is used to decode the
cput value refresh.

Fixes: #143754
Release note: Fixes a bug in the preview logical data replication
feature. Previously, some rows would fail to replicate and go to the DLQ
if the column ids in the source and destination table do not match.