opt: add swap flag to MutationPrivate
Inside the optimizer we can simply use a boolean flag to distinguish
between update and delete, and update swap and delete swap,
respectively. Add this flag to MutationPrivate and use it in execbuild
to construct the swap nodes.

Informs: #144503

Release note: None

opt/norm: add UseSwapMutation normalization rule

Add a new normalization rule which converts Update and Delete
expressions to UpdateSwap and DeleteSwap, respectively. Crucially, this
normalization rule also replace the initial Scan of the mutation with a
single-row Values expression, which optimistically provides the complete
old row (if it exists and matches the filters). This allows the
UpdateSwap and DeleteSwap operators to use CPut to validate the
existence of the old row in the same batch as modifying it, changing the
mutation statement from 2 RTs to 1 RT.

With this normalization rule in place, UpdateSwap and DeleteSwap should
be available for use by any qualifying UPDATE and DELETE statements.

Fixes: #144503
Fixes: #85328

Release note (performance improvement): Add new `update swap` and
`delete swap` operators which allow some UPDATE and DELETE statements to
execute in 1 round trip instead of 2 round trips. For now, these
operators can be used when:

- all columns in the primary index are constrained to a single exact value
  by the WHERE clause
- only a single row is modified
- there are no FK checks or cascades
- there are no uniqueness checks
- there are no check constraints
- there are no after triggers
- the table only uses a single column family

We may lift some of these restrictions in future releases.

logictest: add testcase for incorrect dupe key from UpdateSwap

In #145994 we discovered a problem with batch ordering that could cause
the wrong ConditionFailed error to be returned. This could cause
UpdateSwap and DeleteSwap to incorrectly return a duplicate key error
instead of simply modifying 0 rows.

This problem was fixed by #147117 and #147510. Now that we have
UpdateSwap working, add a testcase that verifies the problem is fixed
for UpdateSwap.

Informs: #144503, #145994

Release note: None

sql, opt: add use_swap_mutations session variable

Release note (sql change): Add new session variable `use_swap_mutations`
which controls whether the new update swap and delete swap operators are
enabled for use by UPDATE and DELETE statements.